src/tesseract/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pytesseract==0.3.10
opencv-python==4.7.0.72
pillow==9.5.0
pdf2image==1.16.3
numpy==1.24.4
python-multipart==0.0.6
aiofiles==23.1.0
typing-extensions==4.7.1
